--- BANK_APPLICATION_DB/BANK_APPLICATION_DB.iml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4" />
