--- BANK_APPLICATION_DB/src/main/resources/hibernate.cfg.xml
@@ -1,36 +1,0 @@
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-
-    <session-factory>
-
-        <!-- JDBC Database connection settings -->
-        <property name="org.hibernate">false</property>
-        <property name="connection.driver_class">com.mysql.jdbc.Driver</property>
-        <property name="connection.url">jdbc:mysql://localhost:3306/BankApplication</property>
-        <property name="connection.username">root</property>
-        <property name="connection.password">root</property>
-
-        <!-- JDBC connection pool settings ... using built-in test pool -->
-        <property name="connection.pool_size">100</property>
-
-
-        <!-- Select our SQL dialect -->
-        <property name="dialect">org.hibernate.dialect.MySQLDialect</property>
-
-        <!-- Echo the SQL to stdout -->
-        <property name="show_sql">true</property>
-
-		<!-- Set the current session context -->
-		<property name="current_session_context_class">thread</property>
-        <property name="hbm2ddl.auto">update</property>
-        <mapping class="com.highpeaksw.customer.BankCustomer"/>
-        <mapping class="com.highpeaksw.accounts.BankAccount"/>
-
- 
-    </session-factory>
-
-</hibernate-configuration>
-
