+++ bankApplicationSpringBoot/src/main/java/META-INF/MANIFEST.MF
@@ -1,0 +1,4 @@
+Manifest-Version: 1.0
+Main-Class: com.highpeaksw.bankApplicationSpringBoot.BankApplicationSp
+ ringBootApplication
+
