+++ bankApplicationSpringBoot/src/main/java/com/highpeaksw/bankApplicationSpringBoot/config/SecurityConfiguration.java
@@ -1,0 +1,24 @@
+package com.highpeaksw.bankApplicationSpringBoot.config;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@EnableWebSecurity
+@Configuration
+public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception{
+        auth.inMemoryAuthentication().withUser("sharky").password("{noop}sharky123").roles("USER").and()
+                .withUser("admin").password("{noop}password").roles("ADMIN");
+    }
+
+    @Override
+    protected void configure(HttpSecurity httpSecurity) throws Exception{
+        httpSecurity.authorizeRequests().anyRequest().fullyAuthenticated().and().httpBasic();
+        httpSecurity.csrf().disable();
+    }
+
+}
