+++ bankApplicationSpringBoot/src/main/java/com/highpeaksw/bankApplicationSpringBoot/repositories/BankCustomerRepositories.java
@@ -1,0 +1,11 @@
+package com.highpeaksw.bankApplicationSpringBoot.repositories;
+
+import com.highpeaksw.bankApplicationSpringBoot.data.BankCustomer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BankCustomerRepositories extends CrudRepository<BankCustomer,Integer> {
+
+}
