+++ bankApplicationSpringBoot/src/main/java/com/highpeaksw/bankApplicationSpringBoot/utils/Utilities.java
@@ -1,4 +1,4 @@
-package com.highpeaksw.utils;
+package com.highpeaksw.bankApplicationSpringBoot.utils;
 
 import java.util.Random;
 import java.util.Scanner;
