+++ bankApplicationSpringBoot/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+#spring.http.encoding.charset=UTF-8
+spring.http.encoding.enabled=true
+server.port=8020
+spring.datasource.driverClassName=com.mysql.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/BankApplication
+spring.datasource.username=root
+spring.datasource.password=root
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.generate-ddl=true
+spring.jpa.show-sql=false
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MariaDB53Dialect
+
+
+
+
