+++ bankApplicationSpringBoot/src/test/java/com/highpeaksw/bankApplicationSpringBoot/BankApplicationSpringBootApplicationTests.java
@@ -1,0 +1,16 @@
+package com.highpeaksw.bankApplicationSpringBoot;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class BankApplicationSpringBootApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
